♻️ restructure noise architecture to allow for better integration of future backend adapters
